Fix issues found by misspell, gofmt -s -w, golint and go vet; add badges
- There were several typos
- Simplified a few simplified statements
- Added doc for a few exported symbols
- Fix incorrect printf usage

+++ host/allwinner/pwm.go
@@ -96,7 +96,7 @@ func (p pwmCtl) String() string {
 	out = append(out, pwmPrescale(p&pwm0PrescaleMask).String())
 	p &^= pwm0PrescaleMask
 	if p != 0 {
-		out = append(out, fmt.Sprintf("Unknown(0x%08X)", p))
+		out = append(out, fmt.Sprintf("Unknown(0x%08X)", uint32(p)))
 	}
 	return strings.Join(out, "|")
 }
